Minor changes to some logic

+++ src/main/kotlin/hr/kotwave/scorpiongym/member/MemberViewModel.kt
@@ -152,8 +152,10 @@ class MemberViewModel(
 
     fun removeMembershipRecord(record: MembershipRecord) {
         _memberRecords.remove(record)
-        membershipRecordDao.deleteAllTrainingsAssociatedWithRecord(record)
-        membershipRecordDao.deleteMembershipRecord(record)
+        if (record.id != 0) {
+            membershipRecordDao.deleteAllTrainingsAssociatedWithRecord(record)
+            membershipRecordDao.deleteMembershipRecord(record)
+        }
     }
 
     fun removeMemberOtherService(memberOtherService: MemberOtherService) {
